// src/app/shared/models/user/user.ts
export class LoginUser {
    username: string;
    password: string;
}

export class CreateUser {
    id?: number;
    username: string;
    firstName: string;
    lastName: string;
    password: string;
    email: string;

}
export class User {
    id: number;
    username: string;
    firstName: string;
    lastName: string;
    password: string;
    email: string;

}